tic-codebase/backend/requirements.txt
asgiref==3.11.0
Django==4.2.26
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.4
typing-extensions==4.15.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.6.0
